PMM-6793 cleanup structure
