+++ pmm-app/src/pmm-add-instance/components/AddRemoteInstance/FormParts/FormParts.messages.ts
@@ -49,5 +49,10 @@ export const Messages = {
       },
       additionalOptions: {},
     },
+    titles: {
+      mainDetails: 'Main details',
+      labels: 'Labels',
+      additionalOptions: 'Additional options',
+    },
   },
 };
