+++ pmm-app/src/pmm-add-instance/components/AddRemoteInstance/FormParts/MainDetails/MainDetails.tsx
@@ -1,0 +1,47 @@
+import React, { FC } from 'react';
+import { PasswordInputField, TextInputField, validators } from '@percona/platform-core';
+import { MainDetailsFormPartProps } from '../FormParts.types';
+import { getStyles } from '../FormParts.styles';
+import { Messages } from '../FormParts.messages';
+import Validators from '../../../../../shared/components/helpers/validators';
+
+export const MainDetailsFormPart: FC<MainDetailsFormPartProps> = ({ remoteInstanceCredentials }) => {
+  const styles = getStyles();
+
+  return (
+    <div className={styles.groupWrapper}>
+      <h5>{Messages.form.titles.mainDetails}</h5>
+      <TextInputField
+        name="address"
+        label={Messages.form.labels.mainDetails.address}
+        placeholder={Messages.form.placeholders.mainDetails.address}
+        validators={[validators.required]}
+        disabled={remoteInstanceCredentials.isRDS}
+      />
+      <TextInputField
+        name="service_name"
+        label={Messages.form.labels.mainDetails.serviceName}
+        placeholder={Messages.form.placeholders.mainDetails.serviceName}
+      />
+      <TextInputField
+        name="port"
+        label={Messages.form.labels.mainDetails.port}
+        placeholder={`Port (default: ${remoteInstanceCredentials.port} )`}
+        validators={[validators.required, Validators.validatePort]}
+      />
+      <TextInputField
+        name="username"
+        label={Messages.form.labels.mainDetails.username}
+        placeholder={Messages.form.placeholders.mainDetails.username}
+        validators={[validators.required]}
+        disabled={remoteInstanceCredentials.isRDS}
+      />
+      <PasswordInputField
+        name="password"
+        label={Messages.form.labels.mainDetails.password}
+        placeholder={Messages.form.placeholders.mainDetails.password}
+        validators={[validators.required]}
+      />
+    </div>
+  );
+};
