+++ pmm-app/src/pmm-add-instance/components/AddRemoteInstance/FormParts/index.tsx
@@ -1,0 +1,3 @@
+export { AdditionalOptionsFormPart as AdditionalOptions } from './AdditionalOptions/AdditionalOptions';
+export { LabelsFormPart as Labels } from './Labels/Labels';
+export { MainDetailsFormPart as MainDetails } from './MainDetails/MainDetails';
